Generate code for integer variable definition
This is the first step towards a different approach of interpreting the Scheme code received as input. The goal is to generate equivalent Java code on the fly, compile and then execute it. This provides the benefit of being able to optimize tail recursion by generating iterative code in the target language. Furthermore, performance should also improve as a result of this.