Add "car" and "cdr" list procedures
It was necessary to add a method to the OutputFormatter which accepts an object of class Object as input and then decides which method to call based upon an "instanceof" check. The reason for this is that the list wrapper only contains objects of class Object, i.e. it has no type information. Thus, "car" can only return an Object instance.