If a procedure is to be called and its name is not known, generate code in case the procedure will be defined later
A procedure "proc1" may refer to another procedure "proc2" which has not yet been defined because the definition of "proc1" precedes the definition of "proc2". In this case, code must be generated (in the form "${procedure_name}(${args})"). If "proc2" is never defined, the program will currently crash at runtime.